[ATOM SGL] runtime extraction
